Bit Changes

+++ app.py
@@ -55,7 +55,7 @@ def news():
 
     return render_template('news.html', newsdb=newsdb)
 
-@app.route('/addNews', methods=['POST'])
+@app.route('/addNews', methods=['GET','POST'])
 def addNews():
     if 'loggedin' in session:
         userId = session['userId']
@@ -68,40 +68,59 @@ def addNews():
             news1 = News(headline=headline, description=description, category=category, author_id=userId)
             db.session.add(news1)
             db.session.commit()
+            flash('News added successfully!', category='info')
             app.logger.info('%s added news' % user.fullName)
             return redirect(url_for('news'))
-    return redirect(url_for('signin'))
+    else:
+        flash('Please Sign-in for adding news!', category='warning')
+        return redirect(url_for('signin'))
+
+    return redirect(url_for('index'))
 
 @app.route('/deleteNews/<int:id>')
 def deleteNews(id):
+    userId = session['userId']
+    news = News.query.get_or_404(id)
+    user = Users.query.get_or_404(userId)
     if 'loggedin' in session:
-        userId = session['userId']
-        user = Users.query.get_or_404(userId)
-        news = News.query.get_or_404(id)
-        db.session.delete(news)
-        db.session.commit()
-        app.logger.info('%s deleted news' % user.fullName)
-        return redirect(url_for('news'))
+        if news.author_id == userId or user.admin:
+            # news = News.query.get_or_404(id)
+            db.session.delete(news)
+            db.session.commit()
+            app.logger.info('%s deleted news' % user.fullName)
+            flash('News deletion successful!', category='info')
+            return redirect(url_for('news'))
+        else:
+            flash('Permission denied!', category='warning')
+            return redirect(url_for('news'))   
     else:
+        flash('Please Sign-in for deleting news!', category='warning')
         return redirect(url_for('signin'))
+
+    return redirect(url_for('index'))
+    
 
 @app.route('/editNews/<int:id>', methods=['GET','POST'])
 def editNews(id):
+    userId = session['userId']
+    news = News.query.get_or_404(id)
+    user = Users.query.get_or_404(userId)
+
     if 'loggedin' in session:
-        userId = session['userId']
-        news = News.query.get_or_404(id)
-        user = Users.query.get_or_404(userId)
-    else:
-        return redirect('signin.html')
-        
-    if request.method == 'POST' and 'headline' in request.form and 'description' in request.form and 'category' in request.form:
-        news.headline = request.form['headline']
-        news.description = request.form['description']
-        news.category = request.form['category']
-        db.session.commit()
-        app.logger.info('%s edited news' % user.fullName)
-        return redirect(url_for('news'))
-        
+        #matching user to his news/post 
+        if news.author_id == userId or user.admin:
+
+            if request.method == 'POST' and 'headline' in request.form and 'description' in request.form and 'category' in request.form:
+                news.headline = request.form['headline']
+                news.description = request.form['description']
+                news.category = request.form['category']
+                db.session.commit()
+                app.logger.info('%s edited news' % user.fullName)
+                return redirect(url_for('news'))
+        else:
+            flash('Permission denied!', category='warning')
+            return redirect(url_for('news'))
+
     return render_template('editNews.html', news=news)
 
 @app.route('/contact')
@@ -117,8 +136,6 @@ def contact():
 def signin():
     if 'loggedin' in session:
         return redirect(url_for('index'))
-    # Output message if something goes wrong...
-    msg = ''
     # Check if "username" and "password" POST requests exist (user submitted form)
     if request.method == 'POST' and 'email' in request.form and 'password' in request.form:
         # Create variables for easy access
@@ -135,23 +152,22 @@ def signin():
                 # Create session data, we can access this data in other routes
                 session['loggedin'] = True
                 session['userId'] = user.id
+                session['admin'] = user.admin
                 # Redirect to home page
                 app.logger.info('%s signed in' % user.fullName)
                 return redirect(url_for('index'))
             else:
-                msg = 'Incorrect password!'
+                flash('Incorrect password!', category='error')
         else:
             # Account doesnt exist or username/password incorrect
-            msg = 'user not found!'
+            flash('user not found!', category='error')
 
-    return render_template('signin.html', msg=msg)
+    return render_template('signin.html')
 
 @app.route('/signup', methods=['GET', 'POST'])
 def signup():
     if 'loggedin' in session:
         return redirect(url_for('index'))
-    # Output message if something goes wrong...
-    msg = ''
     # Check if POST requests exist (user submitted form)
     if request.method == 'POST' and 'fullName' in request.form and 'email' in request.form and 'password' in request.form:
         # Create variables for easy access
@@ -164,9 +180,9 @@ def signup():
         account = Users.query.filter(db.or_(Users.fullName==fullName, Users.email==email)).first()
         # If account exists show error and validation checks
         if account:
-            msg = 'Account already exists!'
+            flash('Account already exists! ', category='error')
         elif not fullName or not password or not email:
-            msg = 'Please fill out the form!'
+            flash('Please fill out the form!', category='error')
         else:
             # Account doesnt exists and the form data is valid, now insert new account into accounts table
             user = Users(fullName=fullName, phone=phone, email=email, password=password)
@@ -174,14 +190,16 @@ def signup():
             db.session.commit()
             session['loggedin'] = True
             session['userId'] = user.id
+            session['admin'] = user.admin
             app.logger.info('%s signed up and logged in' % user.fullName)
+            flash('Registered successfully!', category='info')
             return redirect(url_for('index'))
             
     elif request.method == 'POST':
         # Form is empty... (no POST data)
-        msg = 'Please fill out the form!'
+        flash('Please fill out the form!', category='error')
     # Show signup form with message (if any)
-    return render_template('signup.html', msg=msg)
+    return render_template('signup.html')
 
 @app.route('/profile')
 def profile():
@@ -204,63 +222,131 @@ def logout():
     app.logger.info('user logged out')
     return redirect(url_for('index'))
 
-@app.route('/editProfile', methods=['POST'])
-def editProfile():
-    msg=''
+@app.route('/editProfile/<int:id>', methods=['GET','POST'])
+def editProfile(id):
+    user = Users.query.get_or_404(id)
     if 'loggedin' in session:
-        userId = session['userId']
-        user = Users.query.filter(Users.id==userId).first()
-    
-    if request.method == 'POST' and 'fullName' in request.form and 'email' in request.form:
-        fullName = request.form['fullName']
-        phone = request.form['phone']
-        email = request.form['email']
-        # updating data
-        user.fullName = fullName
-        user.phone = phone
-        user.email = email
-        db.session.commit()
-        app.logger.info('%s updated profile' % user.fullName)
-        return redirect(url_for('profile'))
-
-@app.route('/deleteProfile')
-def deleteProfile():
-    if 'loggedin' in session:
-        userId = session['userId']
-        user = Users.query.filter(Users.id==userId).first()
-        db.session.delete(user)
-        app.logger.info('%s deleted profile' % user.fullName)
-        db.session.commit()
-        logout()
-        return redirect(url_for('index'))
-    else:
-        return redirect(url_for('signin'))
-
-@app.route('/resetPassword', methods=['GET','POST'])
-def resetPassword():
-    msg=''
-    if 'loggedin' in session:
-        userId = session['userId']
-        user = Users.query.filter(Users.id==userId).first()
-    else:
-        return redirect(url_for('signin'))
-
-    if request.method == 'POST' and 'password' in request.form and 'cnfmPassword' in request.form:
-        password = request.form['password']
-        cnfmPassword = request.form['cnfmPassword']
-
-        if password != cnfmPassword:
-            msg ='Passwords do not match! Try again...'
+        if user.id == session['userId'] or session['admin']:
+            if request.method == 'POST' and 'fullName' in request.form and 'email' in request.form:
+                fullName = request.form['fullName']
+                phone = request.form['phone']
+                email = request.form['email']
+                # updating data
+                editUser(user, fullName, phone, email)
+                app.logger.info('%s updated profile' % user.fullName)
+                flash('Profile updated successfully!', category='info')
+                return redirect(url_for('profile'))
         else:
-            pass_hash = bcrypt.generate_password_hash(password).decode('utf-8')
-            user.password = pass_hash
-            db.session.commit()
-            app.logger.info('%s changed password' % user.fullName)
-            flash('Please Sign-in with new password...')
-            logout()
-            return redirect(url_for('signin'))
+            flash('Permission Denied!', category='warning')
+            return redirect(url_for('profile'))
+    else:
+        flash('Please Sign-in first!', category='warning')
+        return redirect(url_for('signin'))
 
-    return render_template('profile.html', user=user, msg=msg)
+@app.route('/deleteProfile/<int:id>')
+def deleteProfile(id):
+    user = Users.query.get_or_404(id)
+
+    if 'loggedin' in session:
+        if user.id == session['userId'] or session['admin']:
+            deleteUser(user)
+            app.logger.info('%s deleted profile' % user.fullName)
+            if session['admin']:
+                return redirect(url_for('ViewUsers'))
+            else:
+                logout()
+                return redirect(url_for('index'))
+        else:
+            flash('Permission Denied!', category='warning')
+            return redirect(url_for('profile'))
+    else:
+        flash('Please Sign-in first!', category='warning')
+        return redirect(url_for('signin'))
+
+@app.route('/resetPassword/<int:id>', methods=['GET','POST'])
+def resetPassword(id):
+    user = Users.query.get_or_404(id)
+    if 'loggedin' in session:
+        if user.id == session['userId'] or session['admin']:
+            if request.method == 'POST' and 'password' in request.form and 'cnfmPassword' in request.form:
+                password = request.form['password']
+                cnfmPassword = request.form['cnfmPassword']
+
+                if password != cnfmPassword:
+                    flash('Passwords do not match! Try again...', category='error')
+                    return redirect(url_for('profile'))
+                else:
+                    resetUserPass(user, password)
+                    app.logger.info('%s changed password' % user.fullName)
+                    flash('Success! Please Sign-in with new password...', category='info')
+                    logout()
+                    return redirect(url_for('signin'))
+        else:
+            flash('Permission Denied!', category='warning')
+            return redirect(url_for('profile'))
+    else:
+        flash('Please Sign-in first!', category='warning')
+        return redirect(url_for('signin'))
+
+# Defining profile handling Functions
+def editUser(user, fullName, phone, email):
+    # updating data
+    user.fullName = fullName
+    user.phone = phone
+    user.email = email
+    db.session.commit()
+
+def deleteUser(user):
+    db.session.delete(user)
+    db.session.commit()
+
+def resetUserPass(user, password):
+    pass_hash = bcrypt.generate_password_hash(password).decode('utf-8')
+    user.password = pass_hash
+    db.session.commit()
+
+@app.route('/viewUsers')
+def viewUsers():
+    if 'loggedin' in session:
+        userId = session['userId']
+        user = Users.query.get_or_404(userId)
+        all_users = Users.query.all()
+
+        return render_template('viewUser.html', all_users = all_users, user=user)
+    else:
+        flash('Please Sign-in first!', category='warning')
+        return redirect(url_for('signin'))
+
+# @app.route('/manageUser/<int:id>', methods=['GET', 'POST'])
+# def manageUser(id):
+#     user = Users.query.get_or_404(id)
+#     if 'loggedin' in session and session['admin']:
+
+#         if request.method == 'POST' and 'fullName' in request.form and 'email' in request.form:
+#             fullName = request.form['fullName']
+#             phone = request.form['phone']
+#             email = request.form['email']
+#             editUser(user, fullName, phone, email)
+#             return redirect(url_for('viewUsers'))
+
+#         elif request.method == 'POST' and 'password' in request.form and 'cnfmPassword' in request.form:
+#             password = request.form['password']
+#             cnfmPassword = request.form['cnfmPassword']
+
+#             if password != cnfmPassword:
+#                 flash('Passwords do not match! Try again...', category='error')
+#                 return redirect(url_for('viewUsers'))            
+#             else:
+#                 resetUserPass(user, password)
+#                 return redirect(url_for('viewUsers'))
+
+#         elif request.method == 'POST' and 'delete' in request.form:
+#             deleteUser(user)
+#             return redirect(url_for('viewUsers'))
+
+#     else:
+#         flash('Permission Denied!', category='warning')
+#         return redirect(url_for('index'))
 
 if __name__== '__main__':
     app.secret_key = 'twhkehberuoraddgcfadsvtw'
